Rename otree app
