+++ binary_choice_game/QnPage.html
