+++ binary_choice_game/utils.py
